Refactored so Heading and TextContent are in components as opposed to styled components within sections
