+++ src/components/Heading/Heading.style.js
@@ -1,0 +1,25 @@
+import styled from "styled-components";
+
+export const StyledLightHeader = styled.h2` {
+    font-size: 1.7rem;
+    font-family: 'Muli', sans-serif;
+    padding: 0 16px 16px;
+    margin: 0px;
+    font-weight: bold;
+    text-align: center;
+    line-height: 1.3;
+    color: #fff;
+}
+`;
+
+export const StyledHeader = styled.h2` {
+    font-size: 1.7rem;
+    font-family: 'Muli', sans-serif;
+    padding: 0 16px 16px;
+    margin: 0px;
+    font-weight: bold;
+    text-align: center;
+    line-height: 1.3;
+    color: #F64C71;
+}
+`;
