+++ src/components/TextContent/TextContent.js
@@ -1,0 +1,13 @@
+import React from "react";
+
+import {StyledText} from "./TextContent.style";
+
+const TextContent = (props) => {
+  return (
+    <StyledText>
+        {props.children}
+    </StyledText>
+  );
+};
+
+export default TextContent;
