+++ src/components/TextContent/TextContent.style.js
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+export const StyledText = styled.p` {
+    font-size: 0.85rem;
+    font-family: 'Muli', sans-serif;
+    padding: 12px 16px;
+    margin: 0px;
+    text-align: center;
+    line-height: 1.3;
+    font-weight: normal;
+  }
+  `;
