+++ src/sections/HowItWorks/HowItWorks.style.js
@@ -3,30 +3,8 @@ import styled from "styled-components";
 export const HowItWorksSection = styled.section` {
     background-color: #F64C71;
     padding: 28px 16px 12px;
+    color: #FFF;
 }
 `;
 
-export const Heading = styled.h2` {
-      font-size: 1.7rem;
-      font-family: 'Muli', sans-serif;
-      padding: 0 16px 16px;
-      font-weight: bold;
-      color: #FFF;
-      text-align: center;
-      line-height: 1.3;
-      margin:0;
-  }
-`;
-
-export const TextContent = styled.p` {
-  font-size: 0.85rem;
-  font-family: 'Muli', sans-serif;
-  padding: 12px 16px;
-  margin: 0px;
-  color: #fff;
-  text-align: center;
-  line-height: 1.3;
-  font-weight: normal;
-}
-`;
