+++ src/sections/PopularRoutes/PopularRoutes.style.js
@@ -5,18 +5,6 @@ export const StyledPopularRoutesSection = styled.section` {
 }
 `;
 
-export const Heading = styled.h2` {
-      font-size: 1.7rem;
-      font-family: 'Muli', sans-serif;
-      padding: 0 16px 16px;
-      margin: 0px;
-      font-weight: bold;
-      color: #F64C71;
-      text-align: center;
-      line-height: 1.3;
-  }
-`;
-
 export const PopularRouteLinks = styled.section` {
   padding: 12px 16px;
   display: flex;
