+++ src/sections/WhatIsWanderWise/WhatIsWanderWise.style.js
@@ -2,30 +2,6 @@ import styled from "styled-components";
 
 export const StyledWhatIsSection = styled.section` {
   padding: 28px 16px 12px;
-}
-`;
-
-export const Heading = styled.h2` {
-      font-size: 1.7rem;
-      font-family: 'Muli', sans-serif;
-      padding: 0 16px 16px;
-      margin: 0px;
-      font-weight: bold;
-      color: #F64C71;
-      text-align: center;
-      line-height: 1.3;
-  }
-`;
-
-export const TextContent = styled.p` {
-  font-size: 0.85rem;
-  font-family: 'Muli', sans-serif;
-  padding: 12px 16px;
-  margin: 0px;
   color: #000;
-  text-align: center;
-  line-height: 1.5;
-  font-weight: normal;
 }
 `;
-
